feat: incorporated shared component into poi
