+++ packages/core/src/sass/account-wizard.scss
@@ -776,3 +776,8 @@
         justify-content: center;
     }
 }
+
+.personal-details-form {
+    width: 100%;
+    padding: 0 16rem;
+}
