+++ packages/core/src/sass/app/_common/components/onfido-container.scss
@@ -75,8 +75,15 @@
         }
 
         .da-inline-note-with-icon {
-            border-radius: 0.8rem;
-            margin-top: 0;
+            // border-radius: 0.8rem;
+            // margin-top: 0;
+            margin-top: 0.8rem;
+            padding: 0.8rem;
+            display: flex;
+            background: rgba(255, 173, 58, 0.16);
+            border-radius: 4px;
+            gap: 0.8rem;
+            line-height: 1.4rem;
         }
 
         .account__scrollbars_container--grid-layout {
@@ -84,8 +91,6 @@
         }
 
         .account-form__section-content {
-            width: 33rem;
-
             .dc-input__bottom-label > * {
                 font-size: var(--text-size-xxs);
             }
@@ -114,6 +119,9 @@
     &__section {
         display: grid;
         grid-template-areas: 'section-side-note section-content';
+        grid-template-columns: 0.5fr 1fr;
+        grid-gap: 1rem;
+        align-items: center;
 
         @include mobile {
             grid-template-areas: 'section-side-note' 'section-content';
@@ -121,7 +129,6 @@
 
         &-side-note {
             grid-area: section-side-note;
-            width: 26rem;
 
             .dc-text {
                 color: var(--text-less-prominent);
@@ -147,7 +154,7 @@
 
         &-content {
             grid-area: section-content;
-            width: 40rem;
+            // width: 40rem;
 
             @include mobile {
                 width: 100%;
